reconstruct agents
